+++ fluidpythran/run.py
@@ -17,6 +17,13 @@ from glob import glob
 from . import __version__
 from .transpiler import make_pythran_files
 from .log import logger, set_log_level
+from .util import compile_pythran_files, ext_suffix, has_to_build
+
+try:
+    import pythran
+except ImportError:
+    pythran = False
+
 
 doc = """
 fluidpythran: easily speedup your Python code with Pythran
@@ -30,6 +37,10 @@ def run():
     See :code:`fluidpythran -h`
     """
     args = parse_args()
+
+    if not args.path:
+        print("No python files given. Nothing to do! ✨ 🍰 ✨.")
+        return
 
     if args.version:
         print(__version__)
@@ -59,6 +70,20 @@ def run():
 
     make_pythran_files(paths, force=args.force)
 
+    if not pythran or args.no_pythran:
+        return
+
+    # find pythran files not already compiled
+    pythran_paths = []
+    for path in paths:
+        path = Path(path)
+        pythran_path = path.parent / "__pythran__" / ("_" + path.name)
+        ext_path = pythran_path.with_suffix(ext_suffix)
+        if has_to_build(ext_path, pythran_path):
+            pythran_paths.append(pythran_path)
+
+    compile_pythran_files(pythran_paths, args.pythran_flags, parallel=True)
+
 
 def parse_args():
     """Parse the arguments"""
@@ -79,5 +104,25 @@ def parse_args():
     )
 
     parser.add_argument("-v", "--verbose", help="verbose mode", action="count")
+
+    parser.add_argument(
+        "-np",
+        "--no-pythran",
+        help="do not compile the Pythran files with Pythran",
+        action="store_true",
+    )
+
+    parser.add_argument(
+        "-pf",
+        "--pythran-flags",
+        help=(
+            "Flags sent to Pythran. "
+            'Default is "-march=native -DUSE_XSIMD". '
+            "There has to be atleast one space in the passed string! "
+            'Example: fluidpythran toto.py -pf "-march=native "\n'
+        ),
+        type=str,
+        default="",
+    )
 
     return parser.parse_args()