+++ CHANGES.rst
@@ -1,3 +1,10 @@
+
+0.0.9 (?)
+---------
+
+- By default, the fluidpythran commandline uses Pythran
+- (NotImplemented) :code:`PYTHRANIZE_AT_IMPORT` mode
+
 0.0.8 (2018-11-16)
 ------------------